Add Vector3D and Tetromino classes for 3D Tetris implementation
